--- src/migration/index.ts
@@ -1,3 +1,0 @@
-export * from './ember-app/index.js';
-export * from './ember-v1-addon/index.js';
-export * from './ember-v2-addon/index.js';
