+++ src/steps/update-project/update-class.ts
